- parser was expecting ints for width and height, needed floats, fixed - parser XmlLevelGoal was incorrectly handling its very ugly color checking section

+++ src/ca/sandstorm/luminance/level/XmlLevel.java
@@ -19,8 +19,8 @@ public class XmlLevel
 	private String _difficulty;
 	private int _xSize;
 	private int _ySize;
-	private int _width;
-	private int _height;
+	private float _width;
+	private float _height;
 	private LinkedList<XmlLevelObject> _objects;
 	private LinkedList<XmlLevelTool> _tools;
 	
@@ -35,7 +35,7 @@ public class XmlLevel
 	 * @precond name != null, difficulty != null, xSize > 0, ySize > 0, objects != null, tools != null
 	 * @postcond An XmlLevel is created
 	 */
-	public XmlLevel(String name, String difficulty, int xSize, int ySize, int width, int height, LinkedList<XmlLevelObject> objects, LinkedList<XmlLevelTool> tools)
+	public XmlLevel(String name, String difficulty, int xSize, int ySize, float width, float height, LinkedList<XmlLevelObject> objects, LinkedList<XmlLevelTool> tools)
 	{
 	    _logger.debug("XmlLevel(" + name + ", " + difficulty + ", " + xSize + ", " + ySize + ", " + objects + ", " + tools + ")");
 
@@ -89,7 +89,7 @@ public class XmlLevel
 	 * Getter method for the width of the level.
 	 * @return The width of the level.
 	 */
-	public int getWidth()
+	public float getWidth()
 	{
 	    return _width;
 	}
@@ -98,7 +98,7 @@ public class XmlLevel
 	 * Getter method for the height of the level.
 	 * @return
 	 */
-	public int getHeight()
+	public float getHeight()
 	{
 	    return _height;
 	}
